Handle sidekiq >= 8 job payload timestamps (#905)
handle sidekiq >= 8 job payload timestamps

fixes #900

In Sidekiq 8, created_at and the like changed from seconds since epoch
to milliseconds since epoch.

https://github.com/sidekiq/sidekiq/blob/main/Changes.md#800

This is handled in sidekiq by switching on the value class (Float vs
Integer) here:
https://github.com/sidekiq/sidekiq/blob/b42be8b804eeeec87bc4b9aa99f4fb62e81d256d/lib/sidekiq/api.rb#L541-L549

In the RubyReaper, the max_score (range end) unix seconds is compared to
an in-progress job's created_at unix milliseconds when checking if a
digest is "active?".

This updates the active? check to convert the timestamp to a seconds
since epoch float before comparing it to max_score (zscore) used to
scroll through digests for reaping.